requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
apache-libcloud==3.8.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.15.0
bcrypt==4.1.3
blinker==1.8.2
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fasteners==0.19
Flask==3.0.3
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Caching==2.3.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlalchemy-file==0.6.0
starlette==0.37.2
starlette-admin==0.13.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
zipp==3.18.2
